python_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
